requirements.txt
anthropic>=0.42.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
python-dotenv>=1.0.0
